dm-crypt: allow trailing "0" argument for compat with Android 5.1.1 vold
The Android 5.1.1 version of vold unconditionally passes an extra "0"
in the dm-crypt options. In newer kernels this is parsed as "there are
no more groups of arguments" and ignored, but in the 3.0-based kernel on
Galaxy Nexus (maguro) it is not accepted.

One approach to this would be to drop the extra "0" from vold, since it
seems to be optional even in newer kernels. However, that would affect
all devices, and I am not able to test that change on a device with a
newer kernel. As a hardware-specific fix for the regression, it seems safer
to add minimal compatibility for the new calling convention in the old kernel.

Thanks to Ivan Zorin for bisecting this.

Bug: https://jira.cyanogenmod.org/browse/NIGHTLIES-1746
Bug: http://forum.xda-developers.com/showpost.php?p=62554113&postcount=3140
Change-Id: Id1761b91861a854982a61609d9705c1da9e7e2c0